requirements.txt
fastapi
uvicorn[standard]
aiofiles
python-multipart
aiohttp
dnspython
python-whois
sqlalchemy
aiosqlite
pydantic
pyyaml
reportlab
